+++ src/module/vue/components/character-sheet-tabs/ironsworn-main.vue
@@ -120,6 +120,6 @@ function assetBrowser() {
       $actor?.toolset ?? 'starforged'
     )
   }
-  theAssetBrowser.render(true)
+  theAssetBrowser.render(true, { focus: true })
 }
 </script>