+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Under the hood: Major styling refactor to replace LESS preprocessor with SCSS/SASS, which supports more robust tooling, including much better support from Stylelint ([#567](https://github.com/ben/foundry-ironsworn/issues/567), [#568](https://github.com/ben/foundry-ironsworn/pull/568))
   - Adds opinionated Stylelint rules for property ordering, variable usage, class naming, etc. We have a _lot_ of styling to manage, and hopefully this will make it a bit easier to maintain.
   - Much of what `palettize.less` did is now written in Typescript for use as a custom SASS plugin. This lets us use `chroma.js` for better colour mixing, and lays the groundwork for generating color themes on the fly.
+- Under the hood: Remove the old `tables.db` (superceded by `ironsworn-oracles.db`) and all remaining references to it ([#614](https://github.com/ben/foundry-ironsworn/pull/614))
+- Add a loading spinner for the asset browser ([#618](https://github.com/ben/foundry-ironsworn/pull/618))
 
 ## 1.20.14